requirements.txt
pandas==2.1.3
black==22.3.0
pre-commit==3.6.0
flake8==6.1.0
hydra-core==1.3.2
numpy==1.26.2
catboost==1.2.2
pyarrow==14.0.1
dvc==3.31.1
dvc-gdrive==2.20.0
joblib==1.3.2
mlflow==2.9.2
sqlalchemy==1.4.39
psycopg2-binary==2.9.9
mlflow==2.9.2
Django==4.2.11
celery[redis]
